Remove dependency on argument-parser
See https://github.com/apple/swift-markdown/pull/149
